培养系统，以及宠物类型的技能盘

+++ haoyou-spring-cloud-alibaba-04-commons-mapper/src/main/java/com/haoyou/spring/cloud/alibaba/commons/entity/LevelUpExp.java
@@ -24,6 +24,6 @@ public class LevelUpExp implements Serializable {
      * 升级所需经验
      */
     @Column(name = "up_lev_exp")
-    private Integer upLevExp;
+    private Long upLevExp;
 
 }
